requirements.txt
clip==1.0
dill==0.3.7
gdown==5.2.0
h5py==3.9.0
hydra-core==1.3.2
lightning==2.0.9.post0
matplotlib==3.7.0
numpy==1.23.5
open_clip_torch==2.24.0
pandas==1.5.3
Pillow==11.1.0
scipy==1.15.2
timm==1.0.15
torchvision==0.16.2
tqdm==4.65.0
